Improved support for local development

+++ docker-setup.sh
@@ -225,7 +225,7 @@ fi
 debug "Finished parameter checks part 1 (@ ${SECONDS})"
 
 docker_setup_tools_file="${docker_setup_cache}/tools.json"
-if test -f "${PWD}/tools.json"; then
+if test "${docker_setup_version}" == "main" && test -f "${PWD}/tools.json"; then
     cp "${PWD}/tools.json" "${docker_setup_cache}/tools.json"
 fi
 if ! test -f "${docker_setup_tools_file}"; then
